explorer: added preset date filtering.. refactored from feeder
